backend/requirements.txt
asgiref==3.8.1
attrs==23.2.0
Django==4.2.8
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
pillow==10.3.0
psycopg==3.1.19
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
typing_extensions==4.11.0
uritemplate==4.1.1
